+++ .settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,3 @@
 eclipse.preferences.version=1
+encoding//src/main/webapp/views/board/reply/reForm.jsp=UTF-8
 encoding/<project>=UTF-8
