+++ src/main/webapp/views/board/ask/askDeleteForm.jsp
@@ -26,7 +26,7 @@ AskboardDTO dto = dao.getAsking(num);
         </script>
     <body>
 <% } else { %>
-     <%@ include file="/views/main/nav.jsp" %>
+     <jsp:include page="/views/main/nav.jsp" />
     
     <br />
     <div class="text-center">
