+++ src/main/webapp/views/board/ask/askForm.jsp
@@ -8,9 +8,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" crossorigin="anonymous"></script>
 </head>
 <%
-try {
 	String memId = (String) session.getAttribute("memId");
-	
 	AskboardDAO dao = new AskboardDAO();
     String nic = dao.select(memId);
     if (memId == null) {
@@ -21,8 +19,14 @@ try {
             window.location="/wherego/views/board/ask/askList.jsp";
         </script>
     <body>
-<% } else { %>
-     <%@ include file="/views/main/nav.jsp" %>
+<% } else { 
+	int num = 0, ref = 1;
+	if(request.getParameter("num")!= null){
+		num=Integer.parseInt(request.getParameter("num"));
+		ref=Integer.parseInt(request.getParameter("ref"));
+	}
+%>
+     <jsp:include page="/views/main/nav.jsp" />
 	
 	<br />
 	<div class="text-center">
@@ -32,7 +36,7 @@ try {
 	<hr />
 	
 	 <!-- 문의 리스트 -->
-    <h2 align="center">문의 게시판</h2>
+    <h2 align="center">Q & A</h2>
     <br />
 	
     <form action="askPro.jsp" method="post" onsubmit="return writeSave()">
@@ -40,26 +44,24 @@ try {
             <label for="exampleFormControlInput1" class="form-label">작성자</label>
             <input type="text" name="writer" class="form-control" id="exampleFormControlInput1" value="<%= nic  %>">
             <input type="hidden" name = id class="form-control" value="<%=memId %>" >
+            <input type="hidden" name = ref class="form-control" value="<%=ref %>" >
         </div>
         <div class="mb-3">
             <label for="exampleFormControlInput2" class="form-label">제목</label>
             <input type="text" name="title" class="form-control" id="exampleFormControlInput2">
         </div>
         <div class="mb-3">
-            <label for="exampleFormControlTextarea1" class="form-label">문의내용</label>
+            <label for="exampleFormControlTextarea1" class="form-label">내용</label>
             <textarea name="content" class="form-control" id="exampleFormControlTextarea1" rows="3"></textarea>
         </div>
         <div class="d-grid gap-2 col-6 mx-auto">
-            <button type="submit" class="btn btn-secondary">문의 등록</button>
+            <button type="submit" class="btn btn-secondary">등록</button>
         </div>
     </form>
     <div class="fixed-bottom">
         <hr />
         <jsp:include page="/views/main/footer.jsp" />
     </div>
-<% }
-} catch (Exception e) {
-}
-%>
+<% }%>
 </body>
 </html>
