+++ src/main/webapp/views/board/ask/askMyList.jsp
@@ -35,7 +35,6 @@
 <head>
 </head>
 <%
-try {
     if (memId == null) {
         
  %>
@@ -46,7 +45,7 @@ try {
     <body>
 <% } else { %>
 <body>
-   <%@ include file="/views/main/nav.jsp" %>
+   <jsp:include page="/views/main/nav.jsp" />
 	
 	<br />
 	<div class="text-center">
@@ -134,11 +133,8 @@ try {
 
     <div >
 	<br/><hr /><br/>
-		<%@ include file="/views/main/footer.jsp" %>	
+		<jsp:include page="/views/main/footer.jsp" />	
 	</div>
-<% }
-} catch (Exception e) {
-}
-%>
+<% }%>
 </body>
 </html>
