+++ src/main/webapp/views/board/notice/noticeUpdateForm.jsp
@@ -15,7 +15,6 @@ int num = Integer.parseInt(request.getParameter("num"));
 NoticeDAO dao = new NoticeDAO();
 NoticeDTO dto = dao.getNoContent(num);
 
-try {
     String memId = (String) session.getAttribute("memId");
     String nic = dao.selectNo(memId);
     if (!memId.equals(dto.getId())) { //등급이 관리자인 사람만 수정 가능 !! 수정하기!!!!
@@ -26,7 +25,7 @@ try {
         </script>
     <body>
 <% } else { %>
-     <%@ include file="/views/main/nav.jsp" %>
+     <jsp:include page="/views/main/nav.jsp" />
     
     <br />
     <div class="text-center">
@@ -62,9 +61,6 @@ try {
         <hr />
         <jsp:include page="/views/main/footer.jsp" />
     </div>
-<% }
-} catch(Exception e) {
-}
-%>
+<% }%>
 </body>
 </html>
